protected routes

+++ src/Quiz/quiz.service.ts
@@ -4,8 +4,11 @@ import {
   HttpStatus,
   Inject,
   Injectable,
+  NotAcceptableException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { response } from 'express';
+import { IAuth } from 'src/Decor/AuthDecor';
 import { Quiz } from 'src/quiz.entity';
 import { Repository } from 'typeorm';
 
@@ -24,29 +27,33 @@ export class QuizService {
     let score = 0;
     correctOrder.forEach((EachQ, index) => {
       let submitted = body[index];
-      if (submitted.length === 1) {
-        let option = EachQ.answerOptions;
-        const ans = option[submitted[0]].isCorrect;
-        if (ans === 'true') {
-          score += 1;
-        }
-      } else if (submitted.length > 1) {
-        submitted = submitted.sort((a: number, b: number) => a - b);
-        let count = 1;
-        EachQ.answerOptions.forEach((option, index) => {
-          if (submitted.includes(index) && option.isCorrect === 'false') {
-            count = 0;
-          }
-        });
-        if (count === 1) {
-          score += 1;
-        }
-      } else {
+      if (submitted?.length === 0) {
         score += 0;
+      } else {
+        if (!EachQ.hasMultiAns) {
+          let option = EachQ.answerOptions;
+          const ans = option[submitted[0]].isCorrect;
+          if (ans === 'true') {
+            score += 1;
+          }
+        } else {
+          submitted = submitted.sort((a: number, b: number) => a - b);
+          let totalAnswer = 0;
+          let count = 0;
+          EachQ.answerOptions.forEach((option, index) => {
+            if (submitted.includes(index) && option.isCorrect === 'true') {
+              count += 1;
+            }
+            if (option.isCorrect === 'true') totalAnswer += 1;
+          });
+          if (count === totalAnswer) {
+            console.log('math');
+            score += 1;
+          }
+        }
       }
     });
 
-    //throw new Error('Method not implemented.');
     return { result: score };
   }
 
@@ -78,23 +85,52 @@ export class QuizService {
   ) {}
   async createQuiz(body: any) {
     const createdQuiz = await this.quizRepository.save(body);
-    // const quiz = Quiz.find({
-    //   where: { title: createdQuiz.title, description: body.description },
-    // });
-
     return createdQuiz;
   }
 
-  async deleteQuiz(id: string) {
+  async deleteQuiz(id: string, auth: IAuth) {
     console.log(id);
+    const Allquiz = await this.quizRepository.find({
+      relations: ['questions', 'user'],
+    });
+    const quiz = Allquiz.find((element) => element.id == id);
+    if (!quiz) {
+      throw new HttpException('Quiz Not Available', HttpStatus.NOT_FOUND);
+    }
+    if (quiz.user.id !== auth.authUser.id) {
+      throw new HttpException('Permission Denied', HttpStatus.NOT_ACCEPTABLE);
+    }
     await this.quizRepository.delete({ id });
     return { response: 'Quiz Deleted' };
   }
 
-  async updatebyID(id: string, body: any) {
+  async updatebyID(id: string, body: any, auth: IAuth) {
+    const oneQuiz = await Quiz.findOne({ where: { id } });
+
+    if (!oneQuiz) {
+      throw new HttpException('Quiz Not Available', HttpStatus.NOT_FOUND);
+    }
+    if (oneQuiz.user.id !== auth.authUser.id) {
+      throw new HttpException('Permission Denied', HttpStatus.NOT_ACCEPTABLE);
+    }
+    if (oneQuiz.isPublished)
+      throw new HttpException('Access Denied', HttpStatus.NOT_ACCEPTABLE);
+    if (body?.isPublished) {
+      const Allquiz = await this.quizRepository.find({
+        where: { id: id },
+        relations: { questions: true },
+      });
+      const quiz = Allquiz.find((element) => element.id == id);
+      if (quiz.questions.length === 0) {
+        throw new HttpException(
+          ' Cannot Publish,No Questions in the Quiz',
+          HttpStatus.NOT_ACCEPTABLE,
+        );
+      }
+    }
     await Quiz.update(id, body);
-    const question = await Quiz.findOneBy({ id });
-    return question;
+    const quiz = await Quiz.findOneBy({ id });
+    return quiz;
   }
 
   async getAll(page: number, limit: number): Promise<any> {
@@ -114,14 +150,28 @@ export class QuizService {
     //return publishedQuizes;
   }
 
-  async getOneById(id: string): Promise<any> {
+  async getOneById(id: string, auth: IAuth): Promise<any> {
     try {
       const Allquiz = await this.quizRepository.find({
-        relations: { questions: true },
+        relations: ['questions', 'user'],
       });
       const quiz = Allquiz.find((element) => element.id == id);
 
-      const { permaLink, isPublished, ...Quiz } = quiz;
+      if (!quiz) {
+        throw new HttpException('Quiz Not Available', HttpStatus.NOT_FOUND);
+      }
+      if (quiz.isPublished)
+        throw new HttpException(
+          'Permission Denied, Quiz Already Published',
+          HttpStatus.NOT_ACCEPTABLE,
+        );
+
+      // console.log('auth', auth);
+      if (quiz.user.id !== auth.authUser.id) {
+        throw new HttpException('Permission Denied', HttpStatus.NOT_ACCEPTABLE);
+      }
+
+      const { permaLink, isPublished, user, ...Quiz } = quiz;
 
       return Quiz;
     } catch (error) {
